scripts/requirements.txt
requests==2.31.0
gspread==6.1.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
